requirements.txt
PyYAML
moviepy
pyinstaller
opencv-python